feat(relayer): make the OpQueue serializable, list operations as json in api
