+++ rust/hyperlane-core/src/chain.rs
@@ -8,6 +8,7 @@ use std::{
 use derive_new::new;
 use num_derive::FromPrimitive;
 use num_traits::FromPrimitive;
+use serde::Serialize;
 #[cfg(feature = "strum")]
 use strum::{EnumIter, EnumString, IntoStaticStr};
 
@@ -39,7 +40,7 @@ impl<'a> std::fmt::Display for ContractLocator<'a> {
 }
 
 /// All domains supported by Hyperlane.
-#[derive(FromPrimitive, PartialEq, Eq, Debug, Clone, Copy, Hash)]
+#[derive(FromPrimitive, PartialEq, Eq, Debug, Clone, Copy, Hash, Serialize)]
 #[cfg_attr(
     feature = "strum",
     derive(strum::Display, EnumString, IntoStaticStr, EnumIter)
@@ -129,7 +130,7 @@ pub enum KnownHyperlaneDomain {
     CosmosTest99991 = 99991,
 }
 
-#[derive(Clone)]
+#[derive(Clone, Serialize)]
 pub enum HyperlaneDomain {
     Known(KnownHyperlaneDomain),
     Unknown {
@@ -155,7 +156,7 @@ impl HyperlaneDomain {
 }
 
 /// Types of Hyperlane domains.
-#[derive(FromPrimitive, Copy, Clone, Eq, PartialEq, Debug)]
+#[derive(FromPrimitive, Copy, Clone, Eq, PartialEq, Debug, Serialize)]
 #[cfg_attr(
     feature = "strum",
     derive(strum::Display, EnumString, IntoStaticStr, EnumIter)
@@ -176,7 +177,7 @@ pub enum HyperlaneDomainType {
 }
 
 /// Hyperlane domain protocol types.
-#[derive(FromPrimitive, Copy, Clone, Eq, PartialEq, Debug)]
+#[derive(FromPrimitive, Copy, Clone, Eq, PartialEq, Debug, Serialize)]
 #[cfg_attr(
     feature = "strum",
     derive(strum::Display, EnumString, IntoStaticStr, EnumIter)
@@ -209,7 +210,7 @@ impl HyperlaneDomainProtocol {
 }
 
 /// Hyperlane domain technical stack types.
-#[derive(Default, FromPrimitive, Copy, Clone, Eq, PartialEq, Debug)]
+#[derive(Default, FromPrimitive, Copy, Clone, Eq, PartialEq, Debug, Serialize)]
 #[cfg_attr(
     feature = "strum",
     derive(strum::Display, EnumString, IntoStaticStr, EnumIter)
